Shifted guidance to utils from protocols as well as guidance_test to tests/utils. Made pull request comment changes also.
