--- demo/guidance_test/tests/misUcount.json
@@ -1,1 +1,0 @@
-{"countData": {"offbyone": 3, "negation": 1, "offbytwo": 3}, "answerDict": {"Test\n\n>>> 1 + 1 # OK will accept 'black' as right answer\ne74918d4310bb6cbc896676f20dc20de\n# locked\n>>> 2 + 2 # OK will accept 'black' as right answer\ne74918d4310bb6cbc896676f20dc20de\n# locked\n": ["['1']", "['0']", "['100']", "['3']", "['4']"]}}
