--- demo/ok_test/ok-test/empty.py
@@ -1,1 +1,0 @@
-# Nothing here for testing
