--- demo/ok_test/ok-test/test.ok
@@ -1,18 +1,0 @@
-{
-    "name": "Lab 6",
-    "endpoint": "cal/cs61a/sp16/test",
-    "src": [
-        "empty.py"
-    ],
-    "tests": {
-        "tests/test.py": "ok_test"
-    },
-    "protocols": [
-        "restore",
-        "file_contents",
-        "unlock",
-        "grading",
-        "analytics",
-        "backup"
-    ]
-}
