--- demo/ok_test/ok-test/tests/simple-test.json
@@ -1,76 +1,0 @@
-{
-  "dictAssessId2WA2DictInfo": {
-    "Test\n\n>>> 1 + 1 # OK will accept 'black' as right answer\ne74918d4310bb6cbc896676f20dc20de\n# locked\n>>> 2 + 2 # OK will accept 'black' as right answer\ne74918d4310bb6cbc896676f20dc20de\n# locked": {
-      "['-2']": {
-        "dictMisU2ReteachMsg": {
-          "negation": 3
-        },
-        "ki": 1,
-        "lstMisU": [
-          "negation"
-        ]
-      },
-      "['0']": {
-        "dictMisU2ReteachMsg": {
-          "negation": 3,
-          "offbytwo": 2
-        },
-        "ki": 1,
-        "lstMisU": [
-          "offbytwo",
-          "negation"
-        ]
-      },
-      "['1']": {
-        "dictMisU2ReteachMsg": {
-          "offbyone": 2
-        },
-        "ki": 1,
-        "lstMisU": [
-          "offbyone"
-        ]
-      },
-      "['100']": {
-        "dictMisU2ReteachMsg": {
-          "offbyone": 2,
-          "offbytwo": 2
-        },
-        "ki": 1,
-        "lstMisU": [
-          "offbyone",
-          "offbytwo"
-        ]
-      },
-      "['3']": {
-        "dictMisU2ReteachMsg": {
-          "offbyone": 2
-        },
-        "ki": 1,
-        "lstMisU": [
-          "offbyone"
-        ]
-      },
-      "['4']": {
-        "dictMisU2ReteachMsg": {
-          "offbytwo": 2
-        },
-        "ki": 1,
-        "lstMisU": [
-          "offbytwo"
-        ]
-      }
-    }
-  },
-  "dictId2Msg": {
-    "1": "KI: Try using your fingers to add",
-    "2": "Reteach: Addition",
-    "3": "Reteach: Negation"
-  },
-  "dictTg2Func": {
-    "0": "lambda info,strMisU: None",
-    "1": "lambda info,strMisU: info['ki']",
-    "2": "lambda info,strMisU: info['dictMisU2ReteachMsg'][strMisU]"
-  },
-  "url": "TODO",
-  "wrongAnsThresh": 2
-}
