--- demo/ok_test/ok-test/tests/tg.ok_tg
@@ -1,1 +1,0 @@
-2
